+++ accounting_reporting_core/src/main/java/org/cardanofoundation/lob/app/accounting_reporting_core/resource/views/ReportView.java
@@ -27,7 +27,7 @@ public class ReportView {
 
     private Optional<Short> period;
 
-    private Integer ver;
+    private Long ver;
 
     private Boolean publish;
